Create hobbies folder and move hobby systems into it
