+++ scripts/globals/festive_moogle.lua
@@ -1,7 +1,7 @@
 -----------------------------------
 -- Festive Moogle
 -----------------------------------
-require('scripts/globals/crafting/crafting_utils')
+require('scripts/globals/hobbies/crafting/utils')
 require('scripts/globals/npc_util')
 -----------------------------------
 xi = xi or {}
