+++ scripts/globals/hobbies/chocobo_digging/data.lua
@@ -1,130 +1,14 @@
 -----------------------------------
--- Chocobo Digging
--- http://ffxiclopedia.wikia.com/wiki/Chocobo_Digging
--- https://www.bg-wiki.com/bg/Category:Chocobo_Digging
------------------------------------
-require('scripts/globals/roe')
-require('scripts/missions/amk/helpers')
+-- Chocobo Digging Data
 -----------------------------------
 xi = xi or {}
 xi.chocoboDig = xi.chocoboDig or {}
-
--- This contais all digging zones with the ones without loot tables defined commented out.
-local diggingZoneList =
-set{
-    xi.zone.CARPENTERS_LANDING,
-    xi.zone.BIBIKI_BAY,
-    -- xi.zone.ULEGUERAND_RANGE,
-    -- xi.zone.ATTOHWA_CHASM,
-    -- xi.zone.LUFAISE_MEADOWS,
-    -- xi.zone.MISAREAUX_COAST,
-    xi.zone.WAJAOM_WOODLANDS,
-    xi.zone.BHAFLAU_THICKETS,
-    -- xi.zone.CAEDARVA_MIRE,
-    -- xi.zone.EAST_RONFAURE_S,
-    -- xi.zone.JUGNER_FOREST_S,
-    -- xi.zone.VUNKERL_INLET_S,
-    -- xi.zone.BATALLIA_DOWNS_S,
-    -- xi.zone.NORTH_GUSTABERG_S,
-    -- xi.zone.GRAUBERG_S,
-    -- xi.zone.PASHHOW_MARSHLANDS_S,
-    -- xi.zone.ROLANBERRY_FIELDS_S,
-    -- xi.zone.WEST_SARUTABARUTA_S,
-    -- xi.zone.FORT_KARUGO_NARUGO_S,
-    -- xi.zone.MERIPHATAUD_MOUNTAINS_S,
-    -- xi.zone.SAUROMUGUE_CHAMPAIGN_S,
-    xi.zone.WEST_RONFAURE,
-    xi.zone.EAST_RONFAURE,
-    xi.zone.LA_THEINE_PLATEAU,
-    xi.zone.VALKURM_DUNES,
-    xi.zone.JUGNER_FOREST,
-    xi.zone.BATALLIA_DOWNS,
-    xi.zone.NORTH_GUSTABERG,
-    xi.zone.SOUTH_GUSTABERG,
-    xi.zone.KONSCHTAT_HIGHLANDS,
-    xi.zone.PASHHOW_MARSHLANDS,
-    xi.zone.ROLANBERRY_FIELDS,
-    -- xi.zone.BEAUCEDINE_GLACIER,
-    -- xi.zone.XARCABARD,
-    -- xi.zone.CAPE_TERIGGAN,
-    xi.zone.EASTERN_ALTEPA_DESERT,
-    xi.zone.WEST_SARUTABARUTA,
-    xi.zone.EAST_SARUTABARUTA,
-    xi.zone.TAHRONGI_CANYON,
-    xi.zone.BUBURIMU_PENINSULA,
-    xi.zone.MERIPHATAUD_MOUNTAINS,
-    xi.zone.SAUROMUGUE_CHAMPAIGN,
-    xi.zone.THE_SANCTUARY_OF_ZITAH,
-    xi.zone.YUHTUNGA_JUNGLE,
-    xi.zone.YHOATOR_JUNGLE,
-    xi.zone.WESTERN_ALTEPA_DESERT,
-    -- xi.zone.QUFIM_ISLAND,
-    -- xi.zone.BEHEMOTHS_DOMINION,
-    -- xi.zone.VALLEY_OF_SORROWS,
-    -- xi.zone.BEAUCEDINE_GLACIER_S,
-    -- xi.zone.XARCABARD_S,
-    -- xi.zone.YAHSE_HUNTING_GROUNDS,
-    -- xi.zone.CEIZAK_BATTLEGROUNDS,
-    -- xi.zone.FORET_DE_HENNETIEL,
-    -- xi.zone.YORCIA_WEALD,
-    -- xi.zone.MORIMAR_BASALT_FIELDS,
-    -- xi.zone.MARJAMI_RAVINE,
-    -- xi.zone.KAMIHR_DRIFTS,
-}
-
-local elementalOreZoneTable =
-set{
-    xi.zone.LA_THEINE_PLATEAU,
-    xi.zone.JUGNER_FOREST,
-    xi.zone.BATALLIA_DOWNS,
-    xi.zone.KONSCHTAT_HIGHLANDS,
-    xi.zone.PASHHOW_MARSHLANDS,
-    xi.zone.ROLANBERRY_FIELDS,
-    xi.zone.TAHRONGI_CANYON,
-    xi.zone.MERIPHATAUD_MOUNTAINS,
-    xi.zone.SAUROMUGUE_CHAMPAIGN,
-}
-
-local diggingWeatherTable =
-{
-    -- Single weather by elemental order.
-    [xi.weather.HOT_SPELL    ] = { xi.item.FIRE_CRYSTAL      },
-    [xi.weather.SNOW         ] = { xi.item.ICE_CRYSTAL       },
-    [xi.weather.WIND         ] = { xi.item.WIND_CRYSTAL      },
-    [xi.weather.DUST_STORM   ] = { xi.item.EARTH_CRYSTAL     },
-    [xi.weather.THUNDER      ] = { xi.item.LIGHTNING_CRYSTAL },
-    [xi.weather.RAIN         ] = { xi.item.WATER_CRYSTAL     },
-    [xi.weather.AURORAS      ] = { xi.item.LIGHT_CRYSTAL     },
-    [xi.weather.GLOOM        ] = { xi.item.DARK_CRYSTAL      },
-
-    -- Double weather by elemental order.
-    [xi.weather.HEAT_WAVE    ] = { xi.item.FIRE_CLUSTER      },
-    [xi.weather.BLIZZARDS    ] = { xi.item.ICE_CLUSTER       },
-    [xi.weather.GALES        ] = { xi.item.WIND_CLUSTER      },
-    [xi.weather.SAND_STORM   ] = { xi.item.EARTH_CLUSTER     },
-    [xi.weather.THUNDERSTORMS] = { xi.item.LIGHTNING_CLUSTER },
-    [xi.weather.SQUALL       ] = { xi.item.WATER_CLUSTER     },
-    [xi.weather.STELLAR_GLARE] = { xi.item.LIGHT_CLUSTER     },
-    [xi.weather.DARKNESS     ] = { xi.item.DARK_CLUSTER      },
-}
-
-local diggingDayTable =
-{
-    [xi.day.FIRESDAY    ] = { xi.item.RED_ROCK,         xi.item.CHUNK_OF_FIRE_ORE      },
-    [xi.day.ICEDAY      ] = { xi.item.TRANSLUCENT_ROCK, xi.item.CHUNK_OF_ICE_ORE       },
-    [xi.day.WINDSDAY    ] = { xi.item.GREEN_ROCK,       xi.item.CHUNK_OF_WIND_ORE      },
-    [xi.day.EARTHSDAY   ] = { xi.item.YELLOW_ROCK,      xi.item.CHUNK_OF_EARTH_ORE     },
-    [xi.day.LIGHTNINGDAY] = { xi.item.PURPLE_ROCK,      xi.item.CHUNK_OF_LIGHTNING_ORE },
-    [xi.day.WATERSDAY   ] = { xi.item.BLUE_ROCK,        xi.item.CHUNK_OF_WATER_ORE     },
-    [xi.day.LIGHTSDAY   ] = { xi.item.WHITE_ROCK,       xi.item.CHUNK_OF_LIGHT_ORE     },
-    [xi.day.DARKSDAY    ] = { xi.item.BLACK_ROCK,       xi.item.CHUNK_OF_DARK_ORE      },
-}
 
 -----------------------------------
 -- Table for common items without special conditions. [Zone ID] = { itemId, weight, dig requirement }
 -- Data from BG wiki: https://www.bg-wiki.com/ffxi/Category:Chocobo_Digging
 -----------------------------------
-local diggingLayer =
+xi.chocoboDig.layer =
 {
     TREASURE = 1, -- This layer takes precedence over all others AND no other layer will trigger if we manage to dig something from it.
     REGULAR  = 2, -- Regular layers. Crystals from weather and ores are applied here.
@@ -136,11 +20,11 @@ xi.chocoboDig.digInfo =
 {
     [xi.zone.CARPENTERS_LANDING] = -- 2
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.KING_TRUFFLE, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.LITTLE_WORM,        100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.ARROWWOOD_LOG,      100, xi.craftRank.AMATEUR    },
@@ -151,7 +35,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.SPRIG_OF_MISTLETOE,  10, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.SCREAM_FUNGUS,       10, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BURROW] = -- Set: Crystals
+        [xi.chocoboDig.layer.BURROW] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -162,7 +46,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.LIGHT_CRYSTAL,     50, xi.craftRank.AMATEUR },
             [8] = { xi.item.DARK_CRYSTAL,      50, xi.craftRank.AMATEUR },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -178,12 +62,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BIBIKI_BAY] = -- 4
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_BIRTH,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.CHUNK_OF_TIN_ORE,       50, xi.craftRank.AMATEUR  },
             [2] = { xi.item.LUGWORM,                50, xi.craftRank.AMATEUR  },
@@ -195,7 +79,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.CHUNK_OF_PLATINUM_ORE,   5, xi.craftRank.ARTISAN  },
             [9] = { xi.item.CORAL_FRAGMENT,          5, xi.craftRank.ARTISAN  },
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -203,7 +87,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -220,15 +104,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.ULEGUERAND_RANGE] = -- 5
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Ores 4
+        [xi.chocoboDig.layer.BURROW] = -- Set: Ores 4
         {
             [1] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_IRON_ORE,         100, xi.craftRank.RECRUIT },
@@ -236,7 +120,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.CHUNK_OF_MYTHRIL_ORE,      100, xi.craftRank.RECRUIT },
             [5] = { xi.item.CHUNK_OF_DARKSTEEL_ORE,     10, xi.craftRank.NOVICE  },
         },
-        [diggingLayer.BORE] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BORE] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -248,22 +132,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.ATTOHWA_CHASM] = -- 7
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] = -- Set: Ores 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 4
         {
             [1] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_IRON_ORE,         100, xi.craftRank.RECRUIT },
@@ -275,19 +159,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.LUFAISE_MEADOWS] = -- 24
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -302,15 +186,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.MISAREAUX_COAST] = -- 25
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BURROW] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -322,7 +206,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.PIECE_OF_ANCIENT_LUMBER,   1, xi.craftRank.CRAFTSMAN  },
             [9] = { xi.item.LACQUER_TREE_LOG,          1, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -330,11 +214,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.WAJAOM_WOODLANDS] = -- 51
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.ALEXANDRITE, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CLUMP_OF_MOKO_GRASS,   100, xi.craftRank.AMATEUR    },
@@ -345,7 +229,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.PEPHREDO_HIVE_CHIP,     10, xi.craftRank.APPRENTICE },
             [8] = { xi.item.CHUNK_OF_ADAMAN_ORE,    10, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BURROW] = -- Set: Logs 2
+        [xi.chocoboDig.layer.BURROW] = -- Set: Logs 2
         {
             [ 1] = { xi.item.CLUMP_OF_MOKO_GRASS,     240, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
@@ -360,7 +244,7 @@ xi.chocoboDig.digInfo =
             [11] = { xi.item.PIECE_OF_ANCIENT_LUMBER,   1, xi.craftRank.CRAFTSMAN  },
             [12] = { xi.item.LACQUER_TREE_LOG,          1, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BORE] = -- Set: Ores 2
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 2
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -374,11 +258,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BHAFLAU_THICKETS] = -- 52
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.ALEXANDRITE, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                  100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.FLINT_STONE,             100, xi.craftRank.AMATEUR    },
@@ -390,7 +274,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.SPIDER_WEB,               10, xi.craftRank.APPRENTICE },
             [9] = { xi.item.CHUNK_OF_ORICHALCUM_ORE,   5, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BURROW] = -- Set: Logs 2
+        [xi.chocoboDig.layer.BURROW] = -- Set: Logs 2
         {
             [ 1] = { xi.item.CLUMP_OF_MOKO_GRASS,     240, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
@@ -405,7 +289,7 @@ xi.chocoboDig.digInfo =
             [11] = { xi.item.PIECE_OF_ANCIENT_LUMBER,   1, xi.craftRank.CRAFTSMAN  },
             [12] = { xi.item.LACQUER_TREE_LOG,          1, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BORE] = -- Set: Ores 2
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 2
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -419,15 +303,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.CAEDARVA_MIRE] = -- 79
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Logs 3
+        [xi.chocoboDig.layer.BURROW] = -- Set: Logs 3
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.DOGWOOD_LOG,             240, xi.craftRank.AMATEUR    },
@@ -438,7 +322,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.EBONY_LOG,                 5, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.PIECE_OF_ANCIENT_LUMBER,   1, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Ores 2
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 2
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -452,22 +336,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.EAST_RONFAURE_S] = -- 81
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Logs 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 4
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR   },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR   },
@@ -483,15 +367,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.JUGNER_FOREST_S] = -- 82
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BURROW] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -502,7 +386,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.REISHI_MUSHROOM,   1, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.KING_TRUFFLE,      1, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Logs 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 4
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR   },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR   },
@@ -518,15 +402,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.VUNKERL_INLET_S] = -- 83
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -534,7 +418,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BORE] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
@@ -545,15 +429,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BATALLIA_DOWNS_S] = -- 84
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Ores 3
+        [xi.chocoboDig.layer.BURROW] = -- Set: Ores 3
         {
             [1] = { xi.item.FLINT_STONE,               240, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
@@ -564,7 +448,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.CHUNK_OF_DARKSTEEL_ORE,     10, xi.craftRank.NOVICE  },
             [8] = { xi.item.CHUNK_OF_SWAMP_ORE,         10, xi.craftRank.NOVICE  },
         },
-        [diggingLayer.BORE] = -- Set: Feathers
+        [xi.chocoboDig.layer.BORE] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -576,15 +460,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.NORTH_GUSTABERG_S] = -- 88
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Ores 3
+        [xi.chocoboDig.layer.BURROW] = -- Set: Ores 3
         {
             [1] = { xi.item.FLINT_STONE,               240, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
@@ -595,7 +479,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.CHUNK_OF_DARKSTEEL_ORE,     10, xi.craftRank.NOVICE  },
             [8] = { xi.item.CHUNK_OF_SWAMP_ORE,         10, xi.craftRank.NOVICE  },
         },
-        [diggingLayer.BORE] = -- Set: Feathers
+        [xi.chocoboDig.layer.BORE] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -607,22 +491,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.GRAUBERG_S] = -- 89
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] = -- Set: Logs 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 4
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR   },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR   },
@@ -638,22 +522,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.PASHHOW_MARSHLANDS_S] = -- 90
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -668,15 +552,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.ROLANBERRY_FIELDS_S] = -- 91
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set:Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set:Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -684,7 +568,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BORE] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -699,22 +583,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.WEST_SARUTABARUTA_S] = -- 95
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -729,19 +613,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.FORT_KARUGO_NARUGO_S] = -- 96
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] = -- Set: Ores 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 1
         {
             [1] = { xi.item.FLINT_STONE,            240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,  100, xi.craftRank.RECRUIT    },
@@ -755,19 +639,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.MERIPHATAUD_MOUNTAINS_S] = -- 97
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -784,15 +668,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.SAUROMUGUE_CHAMPAIGN_S] = -- 98
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -800,7 +684,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Logs 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 4
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR   },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR   },
@@ -816,11 +700,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.WEST_RONFAURE] = -- 100
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.LITTLE_WORM,            50, xi.craftRank.AMATEUR  },
             [ 2] = { xi.item.ACORN,                  50, xi.craftRank.AMATEUR  },
@@ -834,14 +718,14 @@ xi.chocoboDig.digInfo =
             [10] = { xi.item.RONFAURE_CHESTNUT,      10, xi.craftRank.NOVICE   },
             [11] = { xi.item.SPRIG_OF_MISTLETOE,     10, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -856,11 +740,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.EAST_RONFAURE] = -- 101
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.LITTLE_WORM,        50, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.ACORN,              50, xi.craftRank.AMATEUR    },
@@ -874,14 +758,14 @@ xi.chocoboDig.digInfo =
             [10] = { xi.item.SPRIG_OF_MISTLETOE, 10, xi.craftRank.APPRENTICE },
             [11] = { xi.item.KING_TRUFFLE,       10, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -896,12 +780,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.LA_THEINE_PLATEAU] = -- 102
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_GLORY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.PEBBLE,                  50, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.LITTLE_WORM,             50, xi.craftRank.AMATEUR    },
@@ -914,14 +798,14 @@ xi.chocoboDig.digInfo =
             [ 9] = { xi.item.MAHOGANY_LOG,            10, xi.craftRank.NOVICE     },
             [10] = { xi.item.PINCH_OF_DRIED_MARJORAM, 10, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BORE] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
@@ -932,14 +816,14 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.VALKURM_DUNES] = -- 103
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.COIN_OF_DECAY,       5, xi.craftRank.ADEPT },
             [2] = { xi.item.ORDELLE_BRONZEPIECE, 5, xi.craftRank.ADEPT },
             [3] = { xi.item.ONE_BYNE_BILL,       5, xi.craftRank.ADEPT },
             [4] = { xi.item.TUKUKU_WHITESHELL,   5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.LUGWORM,                 50, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BONE_CHIP,              100, xi.craftRank.AMATEUR   },
@@ -951,7 +835,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.SHALL_SHELL,             50, xi.craftRank.CRAFTSMAN },
             [9] = { xi.item.TURTLE_SHELL,            50, xi.craftRank.ARTISAN   },
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -959,7 +843,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -976,13 +860,13 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.JUGNER_FOREST] = -- 104
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_BIRTH,        5, xi.craftRank.ADEPT },
             [3] = { xi.item.KING_TRUFFLE,         5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.LITTLE_WORM,        50, xi.craftRank.AMATEUR   },
             [2] = { xi.item.ACORN,              50, xi.craftRank.AMATEUR   },
@@ -993,7 +877,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.SPRIG_OF_MISTLETOE, 10, xi.craftRank.NOVICE    },
             [8] = { xi.item.SCREAM_FUNGUS,       5, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Crystals
+        [xi.chocoboDig.layer.BURROW] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1004,7 +888,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.LIGHT_CRYSTAL,     50, xi.craftRank.AMATEUR },
             [8] = { xi.item.DARK_CRYSTAL,      50, xi.craftRank.AMATEUR },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -1020,12 +904,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BATALLIA_DOWNS] = -- 105
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_ADVANCEMENT,  5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                50, xi.craftRank.AMATEUR    },
             [2] = { xi.item.FLINT_STONE,           50, xi.craftRank.AMATEUR    },
@@ -1037,7 +921,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.BLACK_CHOCOBO_FEATHER,  5, xi.craftRank.APPRENTICE },
             [9] = { xi.item.REISHI_MUSHROOM,        5, xi.craftRank.JOURNEYMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1045,7 +929,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BORE] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
@@ -1056,12 +940,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.NORTH_GUSTABERG] = -- 106
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_GLORY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.PEBBLE,                 50, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.LITTLE_WORM,            50, xi.craftRank.AMATEUR    },
@@ -1075,14 +959,14 @@ xi.chocoboDig.digInfo =
             [10] = { xi.item.CHUNK_OF_MYTHRIL_ORE,   10, xi.craftRank.APPRENTICE },
             [11] = { xi.item.CHUNK_OF_DARKSTEEL_ORE, 10, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1097,12 +981,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.SOUTH_GUSTABERG] = -- 107
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_DECAY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.PEBBLE,               50, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.LITTLE_WORM,          50, xi.craftRank.AMATEUR    },
@@ -1115,14 +999,14 @@ xi.chocoboDig.digInfo =
             [ 9] = { xi.item.MYTHRIL_BEASTCOIN,    50, xi.craftRank.APPRENTICE },
             [10] = { xi.item.CHUNK_OF_MYTHRIL_ORE, 10, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1137,12 +1021,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.KONSCHTAT_HIGHLANDS] = -- 108
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_BIRTH,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                 50, xi.craftRank.AMATEUR    },
             [2] = { xi.item.FLINT_STONE,            50, xi.craftRank.AMATEUR    },
@@ -1154,7 +1038,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.MYTHRIL_BEASTCOIN,       5, xi.craftRank.APPRENTICE },
             [9] = { xi.item.ELM_LOG,                 5, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Crystals
+        [xi.chocoboDig.layer.BURROW] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1165,7 +1049,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.LIGHT_CRYSTAL,     50, xi.craftRank.AMATEUR },
             [8] = { xi.item.DARK_CRYSTAL,      50, xi.craftRank.AMATEUR },
         },
-        [diggingLayer.BORE] = -- Set: Feathers
+        [xi.chocoboDig.layer.BORE] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1177,12 +1061,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.PASHHOW_MARSHLANDS] = -- 109
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_ADVANCEMENT,  5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,              50, xi.craftRank.AMATEUR    },
             [2] = { xi.item.INSECT_WING,         50, xi.craftRank.AMATEUR    },
@@ -1194,14 +1078,14 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.PETRIFIED_LOG,        5, xi.craftRank.NOVICE     },
             [9] = { xi.item.PUFFBALL,             5, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1216,12 +1100,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.ROLANBERRY_FIELDS] = -- 110
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_GLORY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                  100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.LITTLE_WORM,             100, xi.craftRank.AMATEUR   },
@@ -1233,7 +1117,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.GOLD_BEASTCOIN,            5, xi.craftRank.CRAFTSMAN },
             [9] = { xi.item.CHUNK_OF_ORICHALCUM_ORE,   5, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1241,7 +1125,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BORE] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -1256,19 +1140,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BEAUCEDINE_GLACIER] = -- 111
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -1276,19 +1160,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.XARCABARD] = -- 112
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -1296,15 +1180,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.CAPE_TERIGGAN] = -- 113
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1312,7 +1196,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -1328,11 +1212,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.EASTERN_ALTEPA_DESERT] = -- 114
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.FLINT_STONE,              240, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.BONE_CHIP,                100, xi.craftRank.AMATEUR    },
@@ -1345,7 +1229,7 @@ xi.chocoboDig.digInfo =
             [ 9] = { xi.item.CHUNK_OF_PLATINUM_ORE,     10, xi.craftRank.JOURNEYMAN },
             [10] = { xi.item.PHILOSOPHERS_STONE,         5, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             [1] = { xi.item.BAG_OF_GRAIN_SEEDS,     50, xi.craftRank.AMATEUR    },
             [2] = { xi.item.BAG_OF_VEGETABLE_SEEDS, 50, xi.craftRank.RECRUIT    },
@@ -1355,7 +1239,7 @@ xi.chocoboDig.digInfo =
             [6] = { xi.item.BAG_OF_TREE_CUTTINGS,   10, xi.craftRank.JOURNEYMAN },
             [7] = { xi.item.BAG_OF_CACTUS_STEMS,    10, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Ores 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 1
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -1369,12 +1253,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.WEST_SARUTABARUTA] = -- 115
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_DECAY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,                50, xi.craftRank.AMATEUR  },
             [2] = { xi.item.LITTLE_WORM,           50, xi.craftRank.AMATEUR  },
@@ -1386,14 +1270,14 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.BALL_OF_SARUTA_COTTON, 10, xi.craftRank.INITIATE },
             [9] = { xi.item.ROSEWOOD_LOG,           5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1408,11 +1292,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.EAST_SARUTABARUTA] = -- 116
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [ 1] = { xi.item.PEBBLE,                50, xi.craftRank.AMATEUR  },
             [ 2] = { xi.item.PAPAKA_GRASS,          50, xi.craftRank.AMATEUR  },
@@ -1425,14 +1309,14 @@ xi.chocoboDig.digInfo =
             [ 9] = { xi.item.EBONY_LOG,             10, xi.craftRank.INITIATE },
             [10] = { xi.item.ROSEWOOD_LOG,           5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BURROW] = -- Set: Gysahl Greens
+        [xi.chocoboDig.layer.BURROW] = -- Set: Gysahl Greens
         {
             [1] = { xi.item.BUNCH_OF_GYSAHL_GREENS, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.CHAMOMILE,               50, xi.craftRank.AMATEUR  },
             [3] = { xi.item.GINGER_ROOT,             50, xi.craftRank.RECRUIT  },
             [4] = { xi.item.HEAD_OF_NAPA,            50, xi.craftRank.INITIATE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1447,12 +1331,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.TAHRONGI_CANYON] = -- 117
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_ADVANCEMENT,  5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,            50, xi.craftRank.AMATEUR    },
             [2] = { xi.item.BONE_CHIP,         50, xi.craftRank.AMATEUR    },
@@ -1464,7 +1348,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.GOLD_BEASTCOIN,    10, xi.craftRank.APPRENTICE },
             [9] = { xi.item.CHUNK_OF_GOLD_ORE, 10, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1472,7 +1356,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -1489,7 +1373,7 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BUBURIMU_PENINSULA] = -- 118
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_BIRTH,        5, xi.craftRank.ADEPT },
@@ -1497,7 +1381,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.ONE_BYNE_BILL,        5, xi.craftRank.ADEPT },
             [5] = { xi.item.TUKUKU_WHITESHELL,    5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.LUGWORM,               100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SHELL_BUG,             100, xi.craftRank.AMATEUR    },
@@ -1509,7 +1393,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.CHUNK_OF_PLATINUM_ORE,   5, xi.craftRank.APPRENTICE },
             [9] = { xi.item.CORAL_FRAGMENT,          5, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1517,7 +1401,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -1534,12 +1418,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.MERIPHATAUD_MOUNTAINS] = -- 119
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_ADVANCEMENT,  5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.FLINT_STONE,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.PEBBLE,                100, xi.craftRank.AMATEUR    },
@@ -1550,7 +1434,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.GOLD_BEASTCOIN,         10, xi.craftRank.APPRENTICE },
             [8] = { xi.item.CHUNK_OF_ADAMAN_ORE,     5, xi.craftRank.JOURNEYMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Ores 4
+        [xi.chocoboDig.layer.BURROW] = -- Set: Ores 4
         {
             [1] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_IRON_ORE,         100, xi.craftRank.RECRUIT },
@@ -1558,7 +1442,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.CHUNK_OF_MYTHRIL_ORE,      100, xi.craftRank.RECRUIT },
             [5] = { xi.item.CHUNK_OF_DARKSTEEL_ORE,     10, xi.craftRank.NOVICE  },
         },
-        [diggingLayer.BORE] = -- Set: Ores 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 1
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -1572,12 +1456,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.SAUROMUGUE_CHAMPAIGN] = -- 120
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_GLORY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.FLINT_STONE,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.PEBBLE,                100, xi.craftRank.AMATEUR    },
@@ -1589,7 +1473,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.RED_JAR,                10, xi.craftRank.NOVICE     },
             [9] = { xi.item.GOLD_BEASTCOIN,          5, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1597,7 +1481,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -1614,12 +1498,12 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.THE_SANCTUARY_OF_ZITAH] = -- 121
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
             [2] = { xi.item.COIN_OF_DECAY,        5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.PEBBLE,              100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CLUMP_OF_MOKO_GRASS, 100, xi.craftRank.AMATEUR    },
@@ -1630,7 +1514,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.KING_TRUFFLE,          5, xi.craftRank.NOVICE     },
             [8] = { xi.item.PETRIFIED_LOG,         5, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1638,7 +1522,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -1654,11 +1538,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.YUHTUNGA_JUNGLE] = -- 123
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.BONE_CHIP,              100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.DANCESHROOM,            100, xi.craftRank.AMATEUR    },
@@ -1670,7 +1554,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.KING_TRUFFLE,            10, xi.craftRank.NOVICE     },
             [9] = { xi.item.EBONY_LOG,                5, xi.craftRank.JOURNEYMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BURROW] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -1681,7 +1565,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.REISHI_MUSHROOM,   1, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.KING_TRUFFLE,      1, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -1697,11 +1581,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.YHOATOR_JUNGLE] = -- 124
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.BONE_CHIP,        100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.KAZHAM_PINEAPPLE,  50, xi.craftRank.AMATEUR    },
@@ -1712,7 +1596,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.CORAL_FUNGUS,      10, xi.craftRank.NOVICE     },
             [8] = { xi.item.EBONY_LOG,          5, xi.craftRank.JOURNEYMAN },
         },
-        [diggingLayer.BURROW] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BURROW] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -1723,7 +1607,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.REISHI_MUSHROOM,   1, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.KING_TRUFFLE,      1, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -1739,11 +1623,11 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.WESTERN_ALTEPA_DESERT] = -- 125
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             [1] = { xi.item.PLATE_OF_HEAVY_METAL, 5, xi.craftRank.ADEPT },
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             [1] = { xi.item.BONE_CHIP,              240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.PEBBLE,                 150, xi.craftRank.AMATEUR    },
@@ -1755,7 +1639,7 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.CORAL_FRAGMENT,           5, xi.craftRank.APPRENTICE },
             [9] = { xi.item.PHILOSOPHERS_STONE,       5, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1763,7 +1647,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] = -- Set: Bones
+        [xi.chocoboDig.layer.BORE] = -- Set: Bones
         {
             [ 1] = { xi.item.BONE_CHIP,                 150, xi.craftRank.AMATEUR    },
             [ 2] = { xi.item.HANDFUL_OF_FISH_SCALES,    150, xi.craftRank.AMATEUR    },
@@ -1780,15 +1664,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.QUFIM_ISLAND] = -- 126
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1796,7 +1680,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1811,22 +1695,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BEHEMOTHS_DOMINION] = -- 127
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] = -- Set: Ores 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 1
         {
             [1] = { xi.item.FLINT_STONE,             240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,   100, xi.craftRank.RECRUIT    },
@@ -1840,15 +1724,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.VALLEY_OF_SORROWS] = -- 128
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Feathers
+        [xi.chocoboDig.layer.BURROW] = -- Set: Feathers
         {
             [1] = { xi.item.CLUMP_OF_RED_MOKO_GRASS, 100, xi.craftRank.AMATEUR   },
             [2] = { xi.item.BLACK_CHOCOBO_FEATHER,    50, xi.craftRank.RECRUIT   },
@@ -1856,7 +1740,7 @@ xi.chocoboDig.digInfo =
             [3] = { xi.item.SPIDER_WEB,                5, xi.craftRank.NOVICE    },
             [5] = { xi.item.PHOENIX_FEATHER,           1, xi.craftRank.CRAFTSMAN },
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -1864,15 +1748,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.BEAUCEDINE_GLACIER_S] = -- 136
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set Ores 3
+        [xi.chocoboDig.layer.BURROW] = -- Set Ores 3
         {
             [1] = { xi.item.FLINT_STONE,               240, xi.craftRank.AMATEUR },
             [2] = { xi.item.CHUNK_OF_SILVER_ORE,       100, xi.craftRank.AMATEUR },
@@ -1883,7 +1767,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.CHUNK_OF_DARKSTEEL_ORE,     10, xi.craftRank.NOVICE  },
             [8] = { xi.item.CHUNK_OF_SWAMP_ORE,         10, xi.craftRank.NOVICE  },
         },
-        [diggingLayer.BORE] = -- Set: Crystals
+        [xi.chocoboDig.layer.BORE] = -- Set: Crystals
         {
             [1] = { xi.item.FIRE_CRYSTAL,      50, xi.craftRank.AMATEUR },
             [2] = { xi.item.ICE_CRYSTAL,       50, xi.craftRank.AMATEUR },
@@ -1898,15 +1782,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.XARCABARD_S] = -- 137
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Beastcoins
+        [xi.chocoboDig.layer.BURROW] = -- Set: Beastcoins
         {
             [1] = { xi.item.BEASTCOIN,          100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SILVER_BEASTCOIN,    50, xi.craftRank.RECRUIT    },
@@ -1914,7 +1798,7 @@ xi.chocoboDig.digInfo =
             [4] = { xi.item.MYTHRIL_BEASTCOIN,    5, xi.craftRank.NOVICE     },
             [5] = { xi.item.PLATINUM_BEASTCOIN,   1, xi.craftRank.APPRENTICE },
         },
-        [diggingLayer.BORE] = -- Set: Logs 4
+        [xi.chocoboDig.layer.BORE] = -- Set: Logs 4
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR   },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR   },
@@ -1930,19 +1814,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.YAHSE_HUNTING_GROUNDS] = -- 260
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -1950,19 +1834,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.CEIZAK_BATTLEGROUNDS] = -- 261
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -1970,15 +1854,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.FORET_DE_HENNETIEL] = -- 262
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Shrooms
+        [xi.chocoboDig.layer.BURROW] = -- Set: Shrooms
         {
             [1] = { xi.item.DEATHBALL,       100, xi.craftRank.AMATEUR    },
             [2] = { xi.item.SLEEPSHROOM,     100, xi.craftRank.AMATEUR    },
@@ -1989,7 +1873,7 @@ xi.chocoboDig.digInfo =
             [7] = { xi.item.REISHI_MUSHROOM,   1, xi.craftRank.JOURNEYMAN },
             [8] = { xi.item.KING_TRUFFLE,      1, xi.craftRank.CRAFTSMAN  },
         },
-        [diggingLayer.BORE] = -- Set: Ores 1
+        [xi.chocoboDig.layer.BORE] = -- Set: Ores 1
         {
             [1] = { xi.item.FLINT_STONE,            240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.CHUNK_OF_ALUMINUM_ORE,  100, xi.craftRank.RECRUIT    },
@@ -2003,19 +1887,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.YORCIA_WEALD] = -- 263
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BORE] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
@@ -2026,22 +1910,22 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.MORIMAR_BASALT_FIELDS] = -- 265
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Yellow Ginseng seeds
+        [xi.chocoboDig.layer.BURROW] = -- Set: Yellow Ginseng seeds
         {
             [1] = { xi.item.PIECE_OF_YELLOW_GINSENG, 150, xi.craftRank.AMATEUR  },
             [2] = { xi.item.BAG_OF_WILDGRASS_SEEDS,   50, xi.craftRank.RECRUIT  },
             [3] = { xi.item.BAG_OF_TREE_CUTTINGS,     10, xi.craftRank.INITIATE },
             [4] = { xi.item.BAG_OF_CACTUS_STEMS,       5, xi.craftRank.NOVICE   },
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -2049,19 +1933,19 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.MARJAMI_RAVINE] = -- 266
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] =
+        [xi.chocoboDig.layer.BURROW] =
         {
             -- No entries.
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
@@ -2069,15 +1953,15 @@ xi.chocoboDig.digInfo =
 
     [xi.zone.KAMIHR_DRIFTS] = -- 267
     {
-        [diggingLayer.TREASURE] =
+        [xi.chocoboDig.layer.TREASURE] =
         {
             -- No entries.
         },
-        [diggingLayer.REGULAR] =
+        [xi.chocoboDig.layer.REGULAR] =
         {
             -- No entries.
         },
-        [diggingLayer.BURROW] = -- Set: Logs 1
+        [xi.chocoboDig.layer.BURROW] = -- Set: Logs 1
         {
             [1] = { xi.item.ARROWWOOD_LOG,           240, xi.craftRank.AMATEUR    },
             [2] = { xi.item.YEW_LOG,                 150, xi.craftRank.AMATEUR    },
@@ -2089,319 +1973,9 @@ xi.chocoboDig.digInfo =
             [8] = { xi.item.PIECE_OF_ANCIENT_LUMBER,   1, xi.craftRank.CRAFTSMAN  },
             [9] = { xi.item.LACQUER_TREE_LOG,          1, xi.craftRank.ARTISAN    },
         },
-        [diggingLayer.BORE] =
+        [xi.chocoboDig.layer.BORE] =
         {
             -- No entries.
         },
     },
 }
-
--- This function handles zone and cooldown checks before digging can be attempted, before any animation is sent.
-local function checkDiggingCooldowns(player)
-    -- Check if current zone has digging enabled.
-    local isAllowedZone = diggingZoneList[player:getZoneID()] or false
-
-    if not isAllowedZone then
-        player:messageBasic(xi.msg.basic.WAIT_LONGER, 0, 0)
-
-        return false
-    end
-
-    -- Check digging cooldowns.
-    local currentTime  = GetSystemTime()
-    local skillRank    = player:getSkillRank(xi.skill.DIG)
-    local zoneCooldown = player:getLocalVar('ZoneInTime') + utils.clamp(60 - skillRank * 5, 10, 60)
-    local digCooldown  = player:getLocalVar('[DIG]LastDigTime') + utils.clamp(15 - skillRank * 5, 3, 16)
-
-    if
-        currentTime < zoneCooldown or
-        currentTime < digCooldown
-    then
-        player:messageBasic(xi.msg.basic.WAIT_LONGER, 0, 0)
-
-        return false
-    end
-
-    return true
-end
-
-local function calculateSkillUp(player)
-    local skillRank = player:getSkillRank(xi.skill.DIG)
-    local maxSkill  = utils.clamp((skillRank + 1) * 100, 0, 1000)
-    local realSkill = player:getCharSkillLevel(xi.skill.DIG)
-    local increment = 1
-
-    -- this probably needs correcting
-    local roll = math.random(1, 100)
-
-    -- make sure our skill isn't capped
-    if realSkill < maxSkill then
-        -- can we skill up?
-        if roll <= 15 then
-            if (increment + realSkill) > maxSkill then
-                increment = maxSkill - realSkill
-            end
-
-            -- skill up!
-            player:setSkillLevel(xi.skill.DIG, realSkill + increment)
-
-            -- update the skill rank
-            -- Digging does not have test items, so increment rank once player hits 10.0, 20.0, .. 100.0
-            if (realSkill + increment) >= (skillRank * 100) + 100 then
-                player:setSkillRank(xi.skill.DIG, skillRank + 1)
-            end
-        end
-    end
-end
-
-local function  handleDiggingLayer(player, zoneId, currentLayer)
-    local digTable = xi.chocoboDig.digInfo[zoneId][currentLayer]
-
-    -- Early return.
-    if
-        not digTable or
-        #digTable <= 0
-    then
-        return 0
-    end
-
-    local dTableItemIds  = {}
-    local rewardItem     = 0
-
-    -- Determine moon multiplier.
-    local moon           = VanadielMoonPhase()
-    local rollMultiplier = 1.5 - math.abs(moon - 50) / 50 -- The lower the multiplier, the better for the player.
-    -- Moon phase 0 and 100 -> multiplier = 0.5
-    -- Moon phase 50        -> multiplier = 1.5
-    -- Moon phase 25 and 75 -> multiplier = 1
-
-    -- Add valid items to dynamic table
-    local playerRank = player:getSkillRank(xi.skill.DIG)
-    local randomRoll = 1000
-    local digRate    = 0
-
-    for i = 1, #digTable do
-        randomRoll = utils.clamp(math.floor(math.random(1, 1000) * rollMultiplier), 1, 1000)
-        digRate    = digTable[i][2]
-
-        -- Denim Pants +1 and Black Chocobo Suit
-        if player:getMod(xi.mod.DIG_RARE_ABILITY) > 0 then
-            if digRate >= 100 then
-                digRate = math.floor(digRate / 2)
-            else
-                digRate = digRate * 2
-            end
-        end
-
-        if
-            randomRoll <= digRate and    -- Roll check
-            playerRank >= digTable[i][3] -- Rank check
-        then
-            table.insert(dTableItemIds, #dTableItemIds + 1, digTable[i][1]) -- Insert item ID to table.
-        end
-    end
-
-    -- Add weather crystals and ores to regular layer only.
-    if currentLayer == diggingLayer.REGULAR then
-        local weather            = player:getWeather()
-        local currentDay         = VanadielDayOfTheWeek()
-        local isElementalOreZone = elementalOreZoneTable[player:getZoneID()] or false
-
-        -- Crystals and Clusters.
-        randomRoll = utils.clamp(math.floor(math.random(1, 1000) * rollMultiplier), 1, 1000)
-        if
-            diggingWeatherTable[weather] and
-            randomRoll <= 100
-        then
-            table.insert(dTableItemIds, #dTableItemIds + 1, diggingWeatherTable[weather][1]) -- Insert item ID to table.
-        end
-
-        -- Geodes / Colored Rocks.
-        randomRoll = utils.clamp(math.floor(math.random(1, 1000) * rollMultiplier), 1, 1000)
-        if
-            playerRank >= xi.craftRank.NOVICE and
-            randomRoll <= 50
-        then
-            table.insert(dTableItemIds, #dTableItemIds + 1, diggingDayTable[currentDay][1]) -- Insert item ID to table.
-        end
-
-        -- Elemenal Ores.
-        randomRoll = utils.clamp(math.floor(math.random(1, 1000) * rollMultiplier), 1, 1000)
-        if
-            isElementalOreZone and                                              -- Zone can drop ore.
-            playerRank >= xi.craftRank.CRAFTSMAN and                            -- Digging level must be 60+
-            xi.data.element.getWeatherElement(weather) ~= xi.element.NONE and -- Weather must be elemental.
-            moon >= 7 and moon <= 21 and                                        -- Moon must be between those values.
-            randomRoll <= 100
-        then
-            table.insert(dTableItemIds, #dTableItemIds + 1, diggingDayTable[currentDay][2]) -- Insert item ID to table.
-        end
-    end
-
-    -- Choose a random entry from the valid item table.
-    if #dTableItemIds > 0 then
-        local chosenItem = math.random(1, #dTableItemIds)
-
-        rewardItem = dTableItemIds[chosenItem]
-    end
-
-    return rewardItem
-end
-
-local function handleItemObtained(player, text, itemId)
-    if itemId > 0 then
-        -- Make sure we have enough room for the item.
-        if player:addItem(itemId) then
-            player:messageSpecial(text.ITEM_OBTAINED, itemId)
-        else
-            player:messageSpecial(text.DIG_THROW_AWAY, itemId)
-        end
-    end
-end
-
-local function handleFatigue(player, text, todayDigCount)
-    if math.random(1, 100) <= player:getMod(xi.mod.DIG_BYPASS_FATIGUE) then
-        player:messageSpecial(text.FOUND_ITEM_WITH_EASE)
-    else
-        xi.chocoboDig.updateFatigue(player, todayDigCount + 1)
-    end
-end
-
-xi.chocoboDig.fetchFatigue = function(player)
-    return player:getCharVar('[DIG]DigCount')
-end
-
-xi.chocoboDig.updateFatigue = function(player, newValue)
-    player:setVar('[DIG]DigCount', newValue, NextJstDay())
-end
-
-xi.chocoboDig.start = function(player)
-    local zoneId        = player:getZoneID()
-    local text          = zones[zoneId].text
-    local todayDigCount = xi.chocoboDig.fetchFatigue(player)
-    local currentX      = player:getXPos()
-    local currentZ      = player:getZPos()
-    local currentXSign  = 0
-    local currentZSign  = 0
-
-    if currentX < 0 then
-        currentXSign = 2
-    end
-
-    if currentZ < 0 then
-        currentZSign = 2
-    end
-
-    -----------------------------------
-    -- Early returns and exceptions
-    -----------------------------------
-
-    -- Handle valid zones and digging cooldowns.
-    if not checkDiggingCooldowns(player) then
-        return false -- This means we do not send a digging animation.
-    end
-
-    -- Handle AMK mission 7 (index 6) exception.
-    if
-        xi.settings.main.ENABLE_AMK == 1 and
-        player:getCurrentMission(xi.mission.log_id.AMK) == xi.mission.id.amk.SHOCK_ARRANT_ABUSE_OF_AUTHORITY and
-        xi.amk.helpers.chocoboDig(player, zoneId, text)
-    then
-        -- Note: The helper function handles the messages.
-        player:setLocalVar('[DIG]LastDigTime', GetSystemTime())
-
-        return true
-    end
-
-    -- Handle auto-fail from fatigue.
-    if
-        xi.settings.main.DIG_FATIGUE > 0 and
-        xi.settings.main.DIG_FATIGUE <= todayDigCount
-    then
-        player:messageText(player, text.FIND_NOTHING)
-        player:setLocalVar('[DIG]LastDigTime', GetSystemTime())
-
-        return true
-    end
-
-    -- Handle auto-fail from position.
-    local lastX = player:getLocalVar('[DIG]LastXPos') * (1 - player:getLocalVar('[DIG]LastXPosSign'))
-    local lastZ = player:getLocalVar('[DIG]LastZPos') * (1 - player:getLocalVar('[DIG]LastZPosSign'))
-
-    if
-        currentX >= lastX - 5 and currentX <= lastX + 5 and -- Check current X axis to see if you are too close to your last X.
-        currentZ >= lastZ - 5 and currentZ <= lastZ + 5     -- Check current Z axis to see if you are too close to your last Z.
-    then
-        player:messageText(player, text.FIND_NOTHING)
-        player:setLocalVar('[DIG]LastDigTime', GetSystemTime())
-
-        return true
-    end
-
-    -----------------------------------
-    -- Perform digging
-    -----------------------------------
-
-    -- Set player variables, no matter the result.
-    player:setLocalVar('[DIG]LastXPos', currentX)
-    player:setLocalVar('[DIG]LastZPos', currentZ)
-    player:setLocalVar('[DIG]LastXPosSign', currentXSign)
-    player:setLocalVar('[DIG]LastZPosSign', currentZSign)
-    player:setLocalVar('[DIG]LastDigTime', GetSystemTime())
-
-    -- Handle trasure layer. Incompatible with the other 3 layers. "Early" return.
-    local trasureItemId = handleDiggingLayer(player, zoneId, diggingLayer.TREASURE)
-
-    if trasureItemId > 0 then
-        handleItemObtained(player, text, trasureItemId)
-        handleFatigue(player, text, todayDigCount)
-        calculateSkillUp(player)
-        player:triggerRoeEvent(xi.roeTrigger.CHOCOBO_DIG_SUCCESS)
-
-        return true
-    end
-
-    -- Handle regional currency here. Incompatible with the other 3 layers. "Early" return.
-    -- TODO: Implement logic and message to zones.
-
-    -- Handle regular layer. This also contains, elemental ores, weather crystals and day-element geodes.
-    local regularItemId = handleDiggingLayer(player, zoneId, diggingLayer.REGULAR)
-
-    handleItemObtained(player, text, regularItemId)
-
-    -- Handle Burrow layer. Requires Burrow skill.
-    local burrowItemId = 0
-
-    if xi.settings.main.DIG_GRANT_BURROW > 0 then -- TODO: Implement Chocobo Raising and Burrow chocobo skill. Good luck
-        burrowItemId = handleDiggingLayer(player, zoneId, diggingLayer.BURROW)
-
-        handleItemObtained(player, text, burrowItemId)
-    end
-
-    -- Handle Bore layer. Requires Bore skill.
-    local boreItemId = 0
-
-    if xi.settings.main.DIG_GRANT_BORE > 0 then -- TODO: Implement Chocobo Raising and Bore chocobo skill. Good luck
-        boreItemId = handleDiggingLayer(player, zoneId, diggingLayer.BORE)
-
-        handleItemObtained(player, text, boreItemId)
-    end
-
-    -- Handle skill-up
-    calculateSkillUp(player)
-
-    -- Handle no item OR record of eminence.
-    if
-        regularItemId == 0 and
-        burrowItemId == 0 and
-        boreItemId == 0
-    then
-        player:messageText(player, text.FIND_NOTHING)
-    else
-        handleFatigue(player, text, todayDigCount)
-        player:triggerRoeEvent(xi.roeTrigger.CHOCOBO_DIG_SUCCESS)
-    end
-
-    -- Dig ended. Send digging animation to players.
-    return true
-end
