+++ scripts/globals/hobbies/crafting/ephemeral_moogle.lua
