+++ scripts/globals/hobbies/crafting/utils.lua
